[MOV] spreadsheet: move RecordSelector template to its own file
