+++ addons/spreadsheet/static/src/global_filters/components/filter_value/filter_value.js
@@ -1,6 +1,6 @@
 /** @odoo-module */
 
-import { RecordsSelector } from "../records_selector";
+import { RecordsSelector } from "../records_selector/records_selector";
 import { RELATIVE_DATE_RANGE_TYPES } from "@spreadsheet/helpers/constants";
 import { DateFilterValue } from "../filter_date_value/filter_date_value";
